requirements.txt
Flask==3.1.0
beautifulsoup4==4.13.4
python-dotenv==1.0.1

